backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
aiosqlite
torch==2.2.0
numpy==1.26.3
websockets==12.0.0
tiktoken==0.5.2 
transformers==4.35.0
python-multipart==0.0.6
Pillow==10.1.0
requests>=2.31.0